requirements.txt
pandas==1.0.4
pandas-profiling==2.8.0
tqdm==4.46.1
jinja2==2.11.2
scipy==1.4.1
scikit-learn==0.23.1
scikit-lego==0.5.0
numpy==1.18.4
watermark==2.0.2
seaborn==0.10.1
matplotlib==3.2.1
tensorflow==2.2.0
# tf-models-official==2.2.0
numba==0.49.1
git+https://github.com/gbolmier/funk-svd
selenium==3.141.0
bs4==0.0.1
wordcloud==1.7.0
pillow==7.1.2
nltk==3.5
gensim==3.8.3
git+https://github.com/statisticianinstilettos/recmetrics.git@master#egg=recmetrics
ml-metrics==0.1.4
bidict==0.19.0
